feat: refactor giant components - extract hooks and sub-components (closes #43)
Significantly reduced the size of the two largest components by extracting
reusable hooks and sub-components:

PhaseTimeline.tsx: 1,413 -> 250 lines (82% reduction)
- Extracted usePhaseTimelineData hook for API operations
- Extracted usePhaseValidation hook for date validation logic
- Extracted InlineEdit, PhaseTimelineTable, AddPhaseModal, AddDependencyModal,
  and PhaseTimelineStyles to client/src/components/phase-timeline/

InteractiveTimeline.tsx: 1,223 -> 487 lines (60% reduction)
- Extracted TimelineGrid component for grid and today line rendering
- Extracted DependencyLines component for SVG dependency rendering
- Extracted TimelineItemTooltip component for phase hover tooltips
- Extracted PhaseHandles component for phase boundary handles
- Moved utility functions (calculateItemPosition, calculateDependencyPoints)
  to client/src/components/interactive-timeline/utils.ts

Additional supporting hooks created:
- useTimelineViewport for shared viewport calculation logic
- useVisualPhaseManagerData for reusable data fetching

The remaining components (VisualPhaseManager, ProjectDemandChart, SmartAssignmentModal)
are all under 1,000 lines and already use extracted hooks effectively.

Fixes #43
