+++ client/src/components/interactive-timeline/index.ts
@@ -1,0 +1,7 @@
+export { TimelineGrid } from './TimelineGrid';
+export { DependencyLines } from './DependencyLines';
+export { TimelineItemTooltip } from './TimelineItemTooltip';
+export { PhaseHandles } from './PhaseHandles';
+export type { PhaseHandle } from './PhaseHandles';
+export { calculateItemPosition, calculateDependencyPoints, getLineColor } from './utils';
+export type { ItemPosition, DependencyPoints } from './utils';
