+++ client/src/components/phase-timeline/index.ts
@@ -1,0 +1,5 @@
+export { InlineEdit } from './InlineEdit';
+export { PhaseTimelineTable, getPhaseColor } from './PhaseTimelineTable';
+export { AddPhaseModal } from './AddPhaseModal';
+export { AddDependencyModal } from './AddDependencyModal';
+export { PhaseTimelineStyles } from './PhaseTimelineStyles';
